✅ Minimum Number of Days to Make m Bouquets | Medium | Java | Runtime: 0 ms | Memory: 42.6 MB | 2026-07-06

+++ leetcode/medium/1482-minimum-number-of-days-to-make-m-bouquets/solution.java
@@ -13,7 +13,7 @@ class Solution {
         //m1
         //int ans =-1 ;
         //m2
-        if ((long)k*m > bloomDay.length){
+        if (k*m > bloomDay.length){
             return -1;
         }
